requirements.txt
fastapi
uvicorn[standard]
httpx
beautifulsoup4
lxml
langdetect
python-dotenv
astrapy
